feat: App.tsx 모듈화 및 유지보수성 개선을 위한 리팩토링
App.tsx 파일을 더 작고 집중적인 컴포넌트 및 커스텀 훅으로 분리하여 코드 구성 및 가독성을 향상시켰습니다.

- **`useLightningSale` 훅 추출**: 번개 세일 로직을 `App.tsx`에서 분리했습니다.
- **`useCartActions` 훅 추출**: `handleAdd` 및 `handleCartClick` 함수를 중앙 집중화했습니다.
- **`OrderSummary` 컴포넌트 추출**: 주문 요약 UI 및 관련 계산 로직을 분리했습니다.
- **`LeftContent` 컴포넌트 추출**: 상품 선택 및 장바구니 표시 부분을 분리했습니다.
- 기존 기능 및 상태 관리(`useProductState`, `useCartState`)는 그대로 유지되도록 보장했습니다.
